// src/modules/petition/interface/petition.interface.ts
export enum IPetitionType{
   EXTRAORDINARY_ABILITY = "EXTRAORDINARY_ABILITY",
   NATIONAL_INTEREST_WAIVER = "NATIONAL_INTEREST_WAIVER",
   EXTRAORDINARY_TALENT = "EXTRAORDINARY_TALENT"
}

export enum IPetitionStatus{
    PENDING = "PENDING",
    DECLINED = "DECLINED",
    APPROVED = "APPROVED"
}

export enum IPetitionTimeline{
    REVIEW = "REVIEW",
    COVER_LETTER_AND_CRITERIA_MET = "COVER_LETTER_AND_CRITERIA_MET",
    US_BENEFIT_AND_ENDEAVOR = "US_BENEFIT_AND_ENDEAVOR",
    FINAL_MERIT_AND_CONCLUSION = "FINAL_MERIT_AND_CONCLUSION",
    EXHIBITA_AND_FINAL_REVIEW = "EXHIBITS_AND_FINAL_REVIEW"
}